Implement many generators, fix several perf bugs
Fix a major performance bug in `makeChoice`, where it was incorrectly appending
the choice made to the choice vector even when it got that choice by reading the
vector. This removes a huge amount of the work done.

Make weighted shrink to False
Exit shrinking passes early if we just tried the same choices
Fix an off by 1 error in `replace`
Change from `Storable` to `Unboxed` vector module, which results in the same
perf but indicates that we don't care about passing the values to C