Shrink state refactoring part 2
Convert to State monad

For some reason, the change from `iterate` to `iterateM` caused an infinite
loop, possibly from increased strictness from State. So I replaced it with
`iterateWhileM`, which is essentially `(takeWhile pred) . iterateM` but in a
way that doesn't rely on producing an infinite list lazily.

I think the code overall is better but not hugely so. The need to shift into
State frequently is not amazing.

This code allocates a bit more than before, but I don't see any reason why that
should be, so there are probably easy wins in fixing that.